split conditional flow

+++ survae/data/loaders/cifar10_survae.py
@@ -15,7 +15,7 @@ class Quantize():
         image = image * 255 # [0, 1] -> [0, 255]
         if self.num_bits != 8:
             image = torch.floor(image / 2 ** (8 - self.num_bits)) # [0, 255] -> [0, 2**num_bits - 1]
-        return image.long()
+        return image
 
 class UnsupervisedCIFAR10(CIFAR10):
     def __init__(self, root=root, train=True, transform=None, download=False):
